Added the new Processors agent.
The Processors agent reports details regarding the processing power for
the managed node.

Defined the agent within schema.xml. Wrote the wrapper code to create
and populate an instance of the agent.

The properties for the Processors agent have also been trimmed back to
what is readily available via HAL. A future fix will replace this with
calls to sysfs.