Refactored the Host agent.
Replaced the previous HostWrapper class with the new HostAgent class,
which more closely resembles the style of the ProcessorsAgent class.

Deleted the HostWrapper class entirely.

Refactored the daemon entry point to reduce the number of methods
invoked in order to start or stop the HostAgent instance. Moved the
NICWrapper's setupQMFObject method from private to public so that it can
be invoked by the HostAgent. This is a temporary measure since the
NICWrapper will be replaced by a future patch.