Removes HAL code from the HostAgent codebase.
Now all aspects of the host environment are details by other means. The
fully qualified hostname and architecture are taken from uname(). Memory
is taken from sysinfo(). The machine UUID is taken from DBus rather than
udev since udev reports hardware UUID which is not guaranteed to be
unique, while DBus is designed to be just that.

Refactored, for now, the NICWrapper class to get its own HAL context
rather than it being supplied by the HostAgent class. This way the
HostAgent is completely HAL free.