Moving QMF functionality into a transport layer.
This is the first step towards separating the APIs from the means by
which they are carried between the remote and local system.

A new type, HostListener, defines a type that receives notifications
when the Host fires off an event.

A new type, ProcessorsListener, defines a type that receives
notifications when Processors updates itself.