Moves the CPU properties into the Host API space.
There was no valid reason to keep this set of APIs in their own agent
and class, so those have been eliminated.

A new HostListener method, HostListener::updated(), was added. It is via
this method that the Host will notify all transports that the load
average statistic is updated on the host.