Moved all host-specific data gather to the Platform implementation.
Also removed the C++ Host class and replaced it with C APIs. As a result
I had to remove a lot of other C++ code to make this a strictly C API
set.

Since the previous code was very Linux-specific, it had to move out into
the platform layer to keep the code OS-agnostic.