Gets the CPU model on a Windows machine.
Also introduces a new system for interacting with WMI via external
scripts. A new source module, util.c, contains the method
exec_and_capture_output(char*, char**). It takes as input a command line
to execute and a pointer to a buffer to be allocated and filled with the
output from the command.