Core: Refactor the Agent base class to process QMF requests on the mainloop thread
The normal approach is to have them processed on a special qpid
thread, but this causes big problems when calling fork().

Since it is the active thread that gets duplicated and it knows about
qpid, it also panics when it notices it is no longer connected to the
broker, meaning exec() never gets a chance to run.