Allow cmake to be run by make after a failed attempt.
This change makes a minor tweak to the top level GNUMakefile to allow
"make" to still work after a first attempt resulted in some errors from
cmake (such as missing dependencies).  By not specifying linux.build as
a PHONY target, the commands would not run once the first invokation
created the directory.  Now the commands will always be run when "make"
is executed from the top level directory.

Signed-off-by: Russell Bryant <russell@russellbryant.net>