utilities.h: Make DIMOF() a bit more C++ safe.
Change sizeof(a[0]) to sizeof(0[a]) in the DIM() macro.  This makes the
macro a bit more safe to have around in a C++ code base.  In most cases,
you get the exact same result.  The benefit is that this will blow up if
someone ever accidentally tries to use this on a C++ class that has
overridden the [] operator.

In passing, remove a duplicate definition of DIMOF().

Signed-off-by: Russell Bryant <russell@russellbryant.net>
Acked-by: Andrew Beekhof <andrew@beekhof.net>
(cherry picked from commit 7928985c1b9b7b9d7232612ead5915f3135a4491)