service-qmf: Don't declare the SrvAgent globally.
Global objects in C++ are evil because you can not predict the order
that constructors will fire.  It's best to avoid them whenever possible,
so this patch localizes the instance of SrvAgent to main().

To make this change possible, a void pointer was added to the services
action object that gets passed to the async operation callbacks.  That
way the callbacks had access back into the agent.

Additionally, the callback functions were localized to SrvAgent.  This
allowed raiseEvent() to become a private member function.

Signed-off-by: Russell Bryant <russell@russellbryant.net>
Acked-by: Andrew Beekhof <andrew@beekhof.net>