Create directories before starting broker
The pid and data directories must exist and be writable by the user qpidd
before starting matahari-brokerd. Since these directories may be cleared on
reboot we shouldn't rely on them already existing, and we can't rely on the
broker to create them as qpidd does not have sufficient permissions. So
ensure they are created in the init script.

Signed-off-by: Zane Bitter <zbitter@redhat.com>
Acked-by: Roman Rakus <rrakus@redhat.com>