Implement the identify method for Host agents.
This patch implements the identify method that was defined for the Host
API.  The implementation triggers a system beep for 1 second when the
identify method is invoked.

This implementation is based on documentation that I found for how to
trigger a system beep on Linux and windows.  It still needs to be
tested to make sure it actually works...

Signed-off-by: Russell Bryant <russell@russellbryant.net>
Acked-by: Andrew Beekhof <andrew@beekhof.net>