Some refactoring and bug fixes for the network API.
1) Refactor the network API a bit to avoid exposing any sigar data
structures.  This makes it so the usage of sigar is an internal
implementation detail and is not exposed at all through the API.

2) Fix memory leaks of the interface list.  All users of the function to
get the list of network interfaces leaked some or all of the memory.
This is now cleaned up.  There is documentation on how this cleanup
should be performed in network.h.

3) Fix memory leaks of the MAC address of a network interface.  The API
call returned dynamically allocated memory and the users of the API call
did not clean up.  This function now accepts a buffer to hold the
result, which avoids the need for allocating any memory.

4) Fix memory leaks of the IP address of a network interface.  The same
situation as the MAC address API apply here.

5) Add doxygen comments for various things along the way.

Signed-off-by: Russell Bryant <russell@russellbryant.net>
Acked-by: Andrew Beekhof <andrew@beekhof.net>