Fix services API memory leaks rooted in get_directory_list().
This patch resolves some memory leaks in a few different services API
calls.  All of the leaks are rooted in the usage of
get_directory_list(), since a few of the API calls use this internally.
These API calls returned a GList.  The list and its elements were not
being freed.

The first part of the fix is to ensure that all of the results from
scandir() get freed properly.  To do this, the data for the GList that
gets returned had to be converted to dynamically allocated memory.  The
second part is to go through and add calls to g_list_free_full() on the
results of these API calls.

I also added doxygen documentation for API calls along the way.

Signed-off-by: Russell Bryant <russell@russellbryant.net>