Use DIMOF() to get the length of matahari_options.
This patch converts a number of places that were using the MAX_CHAR
define to using DIMOF(matahari_options).  The result is the exact same
but makes the code a bit more explicit.  It also makes the code slightly
less error prone since none of these places will be subject to an error
being introduced if the array length was changed.

Signed-off-by: Russell Bryant <russell@russellbryant.net>
Acked-by: Andrew Beekhof <andrew@beekhof.net>