Get rid of _instance member in MatahariAgent
This is not actually required by the MatahariAgent superclass. It was acting
as a convenience for subclasses that have exactly one class in the schema.
In the case of schemas with multiple classes, which is the general case, it
didn't make a whole lot of sense. Allow each agent to define its own
instance.

Signed-off-by: Zane Bitter <zbitter@redhat.com>
Acked-by: Russell Bryant <russell@russellbryant.net>