Host: rhbz714249 - Read CPU count from sigar correctly
This patch reads the CPU Core count from sigar correctly. Unfortunately it
does not completely resolve the issue, as sigar is counting each thread in
a HyperThreaded core as a separate core. A change to sigar will be required
to fix that.

Also modified the help for the schema to indicate that cpu_count is, in
fact, returning the logical CPU count (i.e. the number of threads), not the
number of physical CPU sockets.

Signed-off-by: Zane Bitter <zbitter@redhat.com>