Spec: rhbz#731233 - Fix RPM warnings about qpidd user
If the QMF broker is not installed, there may be no qpidd user on the system
(unless qpid is also installed). However, we need to ensure that the qpidd
user can access /var/lib/matahari and /var/run/matahari without changing
the permissions if it is installed at a later time.

The solution is to create the group qpidd if it does not exist and set the
ownership of /var/lib/matahari to root:qpidd with mode 770. This avoids the
need to create the qpidd user's home directory, which should be
/var/lib/qpidd. We leave that to the qpid-cpp-server package.

Note that /var/lib/matahari is no longer world-readable (/var/run/matahari
remains world-readable).

Signed-off-by: Zane Bitter <zbitter@redhat.com>
Acked-by: Russell Bryant <rbryant@redhat.com>