core: rhbz#731858 Fix connection to broker via dns srv.
This bug report pointed out that the agent's log output would show
successfully doing a DNS SRV lookup, but then wouldn't use it.  It would
always first try to connect to localhost.  This patch resolves that.
The fix for the immediate problem is just one line.  The rest is a bit
of an improvement to the DNS SRV functionality.

The only way to use DNS SRV was if you were connecting to the same
domain as what Matahari determines the local machine is a part of (the
result of mh_domainname()).  A new option has been added so that the
domain to connect to can be specified.  For example:

    $ matahari-qmf-hostd -vvv -d russellbryant.net
    mh_connect: SRV record resolved to: justtesting.russellbryant.net
    mh_connect: Trying: amqp:tcp:justtesting.russellbryant.net:49000

Signed-off-by: Russell Bryant <rbryant@redhat.com>