services: rhbz#731914 - Fix crash in services agent.
This patch resolves a crash in the services agent when a requested
action times out.  The problem was that when the action timed out, the
data associated with the action was freed, but the mainloop still had
the stdout and stderr pipes registered for monitoring.  This ensures
that those get unregistered when the action data is being cleaned up.

Signed-off-by: Russell Bryant <rbryant@redhat.com>