qmf: rhbz#732498 Set object names explicitly.
The bug pointed out some inconsistency with object naming.  This patch
changes the QMF agents to ensure that all of them set the object name
explicitly.  This makes object lookup from a console much easier.

The documentation for addData() says that the name must be unique to the
object.  We're hard coding the name, but there is only one instance of
this object per agent, so it's ok.

In passing, the patch also corrects a bug in the sysconfig console.  It
needed to be updated to match the new name of "sysconfig".

Signed-off-by: Russell Bryant <rbryant@redhat.com>