qmf: Don't use string after it has been destroyed.
A number of places grabbed the internal representation of a string and
used it later after the string destructor had run.  In general, the
problems all looked like this:

    const char *foo = args["foo"].asString().c_str();

This is bad.  foo is not valid after this line of code since asString()
is returning a temporary std::string.  This code is grabbing an internal
char * from the string and then it immediately gets destroyed.  Any use
of 'foo' after this is not valid.

Signed-off-by: Russell Bryant <rbryant@redhat.com>