sysconfig: rhbz#733150 set is_postboot_configured property to RW.
This bug reported a sysconfig agent crash due to an unhandled exception
when the is_postboot_configured property was set to an incompatible
value (string instead of an int).

Upon further inspection, the way this property was getting set didn't
make a lot of sense to me.  I proposed removing the property completely
in favor of just using the is_configured method, but beekhof made a good
point that you can't filter on the result of a method call.

Instead of trying to define when is_postboot_configured should be set
inside of Matahari, let that be "business logic" and just get set by the
application using the Matahari APIs.

Signed-off-by: Russell Bryant <rbryant@redhat.com>