core: rhbz#731858 - Use dnsdomainname in automatic SRV lookups.
When an agent is run without a specific broker specified (no --broker
provided), it does some magic to try to figure out where it should
connect.  The first thing it does is try to do an SRV lookup against the
local domain.  The previous code was using the "domainname" as opposed
to the "dnsdomainname", which is more correct for what is being done.

This implementation addresses the issue at hand on Linux.  It leaves the
behavior unchanged on Windows, but I left a note in there as its
possible that there is a similar differentiation on Windows, but that's
not clear to me yet.

Signed-off-by: Russell Bryant <rbryant@redhat.com>