Build: Don't use class name in DBus XML file name
A schema may contain multiple classes, so it makes no sense to name the XML
file generated for DBus after the first class in the file. Instead, just
name it after the package.

However, we have already created multiple daemons in the same package
namespace, so use the existing names when installing.

Signed-off-by: Zane Bitter <zbitter@redhat.com>