services: rhbz#734522 sanity check sevices action parameters.
The bug that was reported showed that if the Resources invoke method was
executed with an empty argument that should be required, it could start
up a recurring action that would fail every time and could never be
cancelled.

Address this by adding some sanity checking on the arguments in the
core services API.  If a parameter is determined to be invalid, don't
attempt the action at all and return an error back to the agent code.
In the QMF agent, if this happens, raise an exception that indicates
that the arguments were invalid.

Signed-off-by: Russell Bryant <rbryant@redhat.com>