Config: rhbz#735429 - Make localhost default broker on Linux
When started with no options, DNS SRV is used (if available) to look up a
broker. In the most common configuration, agents should connect to the
local broker on Linux. The user may then optionally configure broker
federation (when support for that arrives) using the DNS SRV lookup to find
the federated broker.

Therefore, change the default configuration to force connection to
localhost on Linux. The logic remains the same, in that if the broker
option is not defined the DNS SRV will still be queried.

On Windows, where there is no broker installed by default, the behaviour
remains the same as before.

Signed-off-by: Zane Bitter <zbitter@redhat.com>