core: Add API support for multiple DNS SRV records.
This code updates the DNS SRV API to support the fact that there may be
more than one record.  There is still an API call that can be used if
the caller only cares about the top result.  That is what is being used
by MatahariAgent right now.

This patch changes the underlying implementation but leaves the
functionality the same.  It does fix some bugs though, including

 - the previous implementation did not guarantee that the highest
   priority record would be used.
 - the previous code in mh_agent.cpp didn't use the port from SRV at
   all.
 - The windows implementation was totally broken.  Hopefully it's less
   broken now, but it still needs testing.

Signed-off-by: Russell Bryant <rbryant@redhat.com>