Broker: rhbz#735425 - Implement broker federation
Replace the broker wrapper script with a C wrapper that calls execve() on
qpidd. Implement parsing of the MATAHARI_PORT environment variable and the
Broker Federation options in the C wrapper.

Broker federation is configured using the BROKER_FEDERATION environment
variable set in /etc/sysconfig/matahari-broker. The contents of the
variable should be a list of hosts or DNS SRV queries separated by any
combinations of spaces, commas and/or semicolons.

Signed-off-by: Zane Bitter <zbitter@redhat.com>