sysconfig: Pay attention to results from puppet.
The main purpose of this patch is to rework how puppet is executed so
that the result can be obtained and passed back to the caller.  This
patch utilizes infrastructure built into the services library to run
puppet in the background and get an asynchronous callback with the
result once it is completed.

There is another bug fixed in the sysconfig library as a part of this
patch.  Every puppet invokation was leaving a file around on disk.
These files are now cleaned up.

Signed-off-by: Russell Bryant <rbryant@redhat.com>