sysconfig: Support older versions of puppet.
The current sysconfig code was written to require puppet version 2.6.6
or above.  The problem with this is that in Fedora, this version is only
in testing, not stable.  This means installation of Matahari will fail
unless you have the testing repo enabled.

This patch resolves this by checking the installed version of puppet at
runtime and adjusting behavior slightly depending on which version is
found.  The only difference is whether or not we execute "puppet <foo>"
or "puppet apply <foo>".

Signed-off-by: Russell Bryant <rbryant@redhat.com>