sysconfig: Detect more failures with curl.
1) Update the usage of curl to check the HTTP result code before
declaring success.  Previously, a request that would result in
something like a 404 would return success, and puppet would be invoked
on the resulting 404 page content.  This is now caught, an error is
returned, and puppet is not invoked.

2) Fix a curl memory leak that occured when the curl request failed
(for example, because the HTTP connection attempt failed).

3) Add more error checks on the results from curl API calls.

4) Ensure that curl_global_init() is called only once.  The curl docs
say that this function should be called exactly once by a process.

Signed-off-by: Russell Bryant <rbryant@redhat.com>