core: rhbz#739666 - ensure mh_uuid() doesn't return NULL.
QMF has a hissy-fit if you give it NULL when it expects a string.  We
set a UUID as a parameter in all of the agents.  Make the function that
returns the UUID never return NULL.

Ideally, this would never happen.  This functionality depends on the
fact that at least one of systemd or dbus are installed on the system.
A nice enhancement here would be to rework the UUID support so that it's
not split between libmcommon (utilities.c) and libmhost (host.c) and
then make it so that the "Filesystem" UUID is not the only one used
here.  If the "Filesystem" UUID is not found, falling back to the
other types of UUIDs seems like a better thing to do than to just leave
it blank.

Signed-off-by: Russell Bryant <rbryant@redhat.com>