Spec: Fix inverted broker dependencies
The broker should not Require the main matahari package. That is now empty
and files that are required by both the broker and QMF agents are installed
by a subpackage (matahari-core) that provides them.

A large part of the %install stanza is no longer required, since make
install handles it correctly.

Also fixed some rpmlint warnings.

Signed-off-by: Zane Bitter <zbitter@redhat.com>