core: Add verbose option support on Windows.
Add support for the verbose option on Windows.  On Linux, you
increase the verbose level by passing multiple instances of -v.

    $ matahari-qmf-hostd -vvv

On Windows, this option gets set in the registry and can only appear
once.  Make the value from the registry act just like the number of
times -v appears on Linux.

Signed-off-by: Russell Bryant <rbryant@redhat.com>