Build: Install config.h and errors.h
It was previously impossible for an external agent to include the logging.h
header, since that includes utilities.h which includes config.h, which was
not exported. config.h is now installed in include/matahari.

The header agent.h includes definitions for backward-compatibilty that
rely on the inclusion of errors.h.

Signed-off-by: Zane Bitter <zbitter@redhat.com>